+++ src/app/change-sum/change-sum.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import {FormControl} from "@angular/forms";
+import {SettingsService} from "../services/settings-service/settings.service";
 
 @Component({
   selector: 'change-sum',
@@ -6,10 +8,16 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./change-sum.component.css']
 })
 export class ChangeSumComponent implements OnInit {
+  amount = new FormControl(0)
 
-  constructor() { }
+  constructor(private settingsService: SettingsService) { }
 
   ngOnInit(): void {
   }
 
+  onSave() {
+    if (!this.amount.value) return;
+
+    this.settingsService.changeAmount(this.amount.value)
+  }
 }